Added 'from_code' and 'language' options to pass through to xgettext
By default, xgettext interprets input files as ASCII, so in order to use UTF-8 encoded input files, you need to be able to set the `--from-code` option (exposed as `from_code`). In some cases, it is also useful to be able to specify an input language (using `language`, passed through to `--language`) for xgettext.