Allow --add-location parameter
Allow usage of the --add-location parameter. By default, the location
consists of a hash, colon, file name and line number. In more recent
versions of xgettext, this can be altered to 'none' or 'file'. The
default value is 'full', but this is implied when the parameter is
missing.

I want to remove the line numbers from the .pot-file, because they
change frequently and this causes differences in the .pot-file that are
not really changes, but do show op in the versioning system each time.